requirements.txt
Django==1.6.6
PyYAML==3.11
boto==2.32.1
dj-database-url==0.3.0
dj-static==0.0.6
django-ckeditor==4.0.2
django-filter==0.7
django-mptt==0.6.1
django-storages==1.1.8
django-tag-parser==1.1.0
django-toolbelt==0.0.1
djangorestframework==2.3.14
gunicorn==19.1.1
psycopg2==2.5.3
static3==0.5.1
wsgiref==0.1.2
